Add Duration#to_rounded_s method
This method converts a Duration to a human-readable string using a rounded value.

By 'rounded', we mean that the resulting value is rounded up if the input includes a value of more than half of one of the least-significant unit to be returned. For example, `(17.hours 43.minutes 31.seconds)`, when rounded to two units (hours and minutes), would return "17 hours, 44 minutes". By contrast, `#to_s`, with a `:count` option of 2, would return a value of "17 hours, 43 minutes": truncating, rather than rounding.

Note that this method overloads the meaning of the `:count` option value as documented. If the passed-in option value is numeric, it will be honored, and rounding will take place to that number of units. If the value is either `:all` or the default `nil`, then _rounding_ will be done to two units, and the rounded value will be passed on to `#to_s` with the options specified (which will result in a maximum of two time units being output).

The passed-in options are passed to `#to_s` as specified; rounding the _value_ to be formatted does not affect its _formatting_.